+++ src/js/data/2025-04-03.js
@@ -6230,7 +6230,7 @@ dataSet[dataSetVersion].characterData = [
   },
   {
     name: "Vainamoinen",
-    img: "unreleased/icon_vainamoinen01.png",
+    img: "3-star/icon_vainamoinen01.png",
     opts: {
       rarity: [ "3+" ],
 	  guild: [ "creators" ],