+++ src/js/data/2025-11-29.js
@@ -6593,7 +6593,7 @@ dataSet[dataSetVersion].characterData = [
   },
   {
     name: "Vainamoinen",
-    img: "unreleased/icon_vainamoinen01.png",
+    img: "3-star/icon_vainamoinen01.png",
     opts: {
       rarity: [ "3+" ],
 	  guild: [ "creators" ],
@@ -6617,7 +6617,7 @@ dataSet[dataSetVersion].characterData = [
   },
   {
     name: "Koshchei",
-    img: "unreleased/icon_koshchei01.png",
+    img: "3-star/icon_koshchei01.png",
     opts: {
       rarity: [ "3+" ],
 	  guild: [ "agents" ],
@@ -6629,7 +6629,7 @@ dataSet[dataSetVersion].characterData = [
   },
   {
     name: "Red Hood",
-    img: "unreleased/icon_redhood01.png",
+    img: "3-star/icon_redhood01.png",
     opts: {
       rarity: [ "3+" ],
 	  guild: [ "no_guild" ],
@@ -6641,7 +6641,7 @@ dataSet[dataSetVersion].characterData = [
   },
   {
     name: "Shisa",
-    img: "unreleased/icon_shisa01.png",
+    img: "3-star/icon_shisa01.png",
     opts: {
       rarity: [ "3+" ],
 	  guild: [ "no_guild" ],
@@ -6665,7 +6665,7 @@ dataSet[dataSetVersion].characterData = [
   },
   {
     name: "Monaciello",
-    img: "unreleased/icon_monaciello01.png",
+    img: "3-star/icon_monaciello01.png",
     opts: {
       rarity: [ "3+" ],
 	  guild: [ "no_guild" ],
@@ -6989,7 +6989,7 @@ dataSet[dataSetVersion].characterData = [
   },
     {
     name: "Red Extra",
-    img: "unreleased/icon_extrafire01b.png",
+    img: "2-star/icon_extrafire01b.png",
     opts: {
       rarity: [ "2" ],
 	  guild: [ "summoners" ],
@@ -7001,7 +7001,7 @@ dataSet[dataSetVersion].characterData = [
   },
   {
     name: "Green Extra",
-    img: "unreleased/icon_extraearth01b.png",
+    img: "2-star/icon_extraearth01b.png",
     opts: {
       rarity: [ "2" ],
 	  guild: [ "summoners" ],
@@ -7013,7 +7013,7 @@ dataSet[dataSetVersion].characterData = [
   },
   {
     name: "Purple Extra",
-    img: "unreleased/icon_extradark01b.png",
+    img: "2-star/icon_extradark01b.png",
     opts: {
       rarity: [ "2" ],
 	  guild: [ "summoners" ],
@@ -7025,7 +7025,7 @@ dataSet[dataSetVersion].characterData = [
   },
   {
     name: "Brown Extra",
-    img: "unreleased/icon_extraworld01b.png",
+    img: "2-star/icon_extraworld01b.png",
     opts: {
       rarity: [ "2" ],
 	  guild: [ "summoners" ],